requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
scikit-learn>=1.4.0
lightgbm>=4.3.0
joblib>=1.3.2
requests>=2.31.0
earthaccess>=0.11.0
s3fs>=2024.6.0
h5py>=3.11.0
